Refactor SVGs to components to direcly use as SVG tags, svelte is a bit weird with svgs :/
