+++ frontend/src/lib/assets/Logo.svelte
