+++ frontend/src/lib/assets/bitcoin.svelte
