+++ frontend/src/lib/assets/chainlink.svelte
