+++ frontend/src/lib/assets/ethereum.svelte
