+++ frontend/src/lib/assets/filecoin.svelte
