+++ frontend/src/lib/components/HeroSection.svelte
@@ -1,26 +1,17 @@
-
-
-
 <script lang="ts">
-	
 	import { balanceOnBlock } from '$lib/stores/state';
 	import { formatEther } from 'ethers/lib/utils';
-  import Logo from '../../../static/Logo.svg'
-
-
+	// import Logo from '$lib/assets/Logo.svg';
 </script>
 
-
-
 <section class="hero">
-  <div class="hero-container">
-      <div class="left">
-          <h1>This is a large header</h1>
-
-      </div>
-      <div class="right">
-          <!-- <CardCta /> -->
-          <h1>anothe component</h1>
-      </div>
-  </div>
+	<div class="hero-container">
+		<div class="left">
+			<h1>This is a large header</h1>
+		</div>
+		<div class="right">
+			<!-- <CardCta /> -->
+			<h1>anothe component</h1>
+		</div>
+	</div>
 </section>
