+++ frontend/svelte.config.js
@@ -3,38 +3,38 @@ import preprocess from 'svelte-preprocess';
 import svg from '@poppanator/sveltekit-svg';
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	// Consult https://github.com/sveltejs/svelte-preprocess
-	// for more information about preprocessors
-	preprocess: [
-		preprocess({
-			postcss: true
-		})
-	],
+  // Consult https://github.com/sveltejs/svelte-preprocess
+  // for more information about preprocessors
+  preprocess: [
+    preprocess({
+      postcss: true
+    })
+  ],
 
-	kit: {
-		adapter: adapter({
-			pages: 'build',
-			assets: 'build',
-			fallback: null,
-			precompress: false
-		}),
-		prerender: {
-			enabled: true,
-			default: true
-		},
-		files: { lib: 'src/lib', assets: 'static' },
+  kit: {
+    adapter: adapter({
+      pages: 'build',
+      assets: 'build',
+      fallback: null,
+      precompress: false
+    }),
+    prerender: {
+      enabled: true,
+      default: true
+    },
+    files: { lib: 'src/lib', assets: 'static' },
     vite: {
       // Options are optional
       plugins: [svg({
-		includePaths: ["./static/"],
-		svgoOptions: {
-		  multipass: true,
-		  plugins: ["preset-default" ],
-		},
-	  } )]
+        includePaths: ["./static/"],
+        svgoOptions: {
+          multipass: true,
+          plugins: ["preset-default"],
+        },
+      })]
     }
-	},
-	
+  },
+
 };
 
 export default config;
